Changed how raw mouse wheel delta is handled.

+++ src/framework/input_system.tcc
@@ -161,9 +161,9 @@ void InputSystem<InputContext>::Update()
 				}
 				break;
 			case RawInputAxis::MouseScroll:
-				if (m_rawInputHandler.IsMouseInWindow() && m_rawInputHandler.GetMouseWheelDelta() != 0.0f)
+				if (m_rawInputHandler.IsMouseInWindow() && m_rawInputHandler.GetMouseWheelDelta() != 0)
 				{
-					m_activeRanges.emplace(range, m_rawInputHandler.GetMouseWheelDelta());
+					m_activeRanges.emplace(range, static_cast<float>(m_rawInputHandler.GetMouseWheelDelta()));
 				}
 				break;
 			}
